refactor: remove /new/ directory
