+++ src/lib/characterPreview/BuffsAnalysisDisplay.tsx
@@ -9,7 +9,7 @@ import { formatOptimizerDisplayData } from 'lib/optimization/optimizer'
 import { Assets } from 'lib/rendering/assets'
 import { originalScoringParams, SimulationScore } from 'lib/scoring/simScoringUtils'
 import { aggregateCombatBuffs } from 'lib/simulations/combatBuffsAnalysis'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
 import { cardShadow } from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
 import { currentLocale } from 'lib/utils/i18nUtils'
 import { TsUtils } from 'lib/utils/TsUtils'
