+++ src/lib/characterPreview/CharacterScoringSummary.tsx
@@ -14,7 +14,7 @@ import { toBasicStatsObject } from 'lib/optimization/basicStatsArray'
 import { Key, StatToKey, toComputedStatsObject } from 'lib/optimization/computedStatsArray'
 import { Assets } from 'lib/rendering/assets'
 import { diminishingReturnsFormula, ScoringType, SimulationScore, spdDiminishingReturnsFormula } from 'lib/scoring/simScoringUtils'
-import { Simulation } from 'lib/simulations/new/statSimulationTypes'
+import { Simulation } from 'lib/simulations/statSimulationTypes'
 import DB from 'lib/state/db'
 import { ColorizedLinkWithIcon } from 'lib/ui/ColorizedLink'
 import { VerticalDivider } from 'lib/ui/Dividers'
