+++ src/lib/characterPreview/characterPreviewController.tsx
@@ -12,8 +12,8 @@ import { RelicScorer, RelicScoringResult } from 'lib/relics/relicScorerPotential
 import { Assets } from 'lib/rendering/assets'
 import { AsyncSimScoringExecution } from 'lib/scoring/dpsScore'
 import { ScoringType } from 'lib/scoring/simScoringUtils'
-import { simulateBuild } from 'lib/simulations/new/simulateBuild'
-import { SimulationRelicByPart } from 'lib/simulations/new/statSimulationTypes'
+import { simulateBuild } from 'lib/simulations/simulateBuild'
+import { SimulationRelicByPart } from 'lib/simulations/statSimulationTypes'
 import { DB } from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
