+++ src/lib/characterPreview/summary/DpsScoreMainStatUpgradesTable.tsx
@@ -4,7 +4,7 @@ import { MainStats, Parts } from 'lib/constants/constants'
 import { iconSize } from 'lib/constants/constantsUi'
 import { Assets } from 'lib/rendering/assets'
 import { SimulationScore } from 'lib/scoring/simScoringUtils'
-import { SimulationStatUpgrade } from 'lib/simulations/new/scoringUpgrades'
+import { SimulationStatUpgrade } from 'lib/simulations/scoringUpgrades'
 import { arrowColor, arrowDirection } from 'lib/tabs/tabOptimizer/analysis/StatsDiffCard'
 import { cardShadowNonInset } from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
 import { localeNumber_0, localeNumber_00 } from 'lib/utils/i18nUtils'
