+++ src/lib/characterPreview/summary/SubstatRollsSummary.tsx
@@ -2,7 +2,7 @@ import { Flex } from 'antd'
 import { Stats, SubStats } from 'lib/constants/constants'
 import { defaultGap } from 'lib/constants/constantsUi'
 import { diminishingReturnsFormula, spdDiminishingReturnsFormula } from 'lib/scoring/simScoringUtils'
-import { SimulationRequest } from 'lib/simulations/new/statSimulationTypes'
+import { SimulationRequest } from 'lib/simulations/statSimulationTypes'
 import { VerticalDivider } from 'lib/ui/Dividers'
 import { numberToLocaleString } from 'lib/utils/i18nUtils'
 import React from 'react'
