+++ src/lib/gpu/tests/webgpuTestGenerator.ts
@@ -4,7 +4,7 @@ import { getWebgpuDevice } from 'lib/gpu/webgpuDevice'
 import { RelicsByPart } from 'lib/gpu/webgpuTypes'
 import { SortOption } from 'lib/optimization/sortOptions'
 
-import { generateFullDefaultForm } from 'lib/simulations/new/utils/benchmarkForm'
+import { generateFullDefaultForm } from 'lib/simulations/utils/benchmarkForm'
 import DB from 'lib/state/db'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { Form } from 'types/form'
