+++ src/lib/gpu/webgpuOptimizer.ts
@@ -9,9 +9,9 @@ import { OptimizerDisplayData } from 'lib/optimization/bufferPacker'
 import { ComputedStatsArray, ComputedStatsArrayCore } from 'lib/optimization/computedStatsArray'
 import { formatOptimizerDisplayData } from 'lib/optimization/optimizer'
 import { SortOption } from 'lib/optimization/sortOptions'
-import { simulateBuild } from 'lib/simulations/new/simulateBuild'
-import { SimulationRelicByPart } from 'lib/simulations/new/statSimulationTypes'
-import { initializeContextConditionals } from 'lib/simulations/new/workerContextTransform'
+import { initializeContextConditionals } from 'lib/simulations/contextConditionals'
+import { simulateBuild } from 'lib/simulations/simulateBuild'
+import { SimulationRelicByPart } from 'lib/simulations/statSimulationTypes'
 import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { activateZeroResultSuggestionsModal } from 'lib/tabs/tabOptimizer/OptimizerSuggestionsModal'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
