+++ src/lib/optimization/context/calculateContext.ts
@@ -3,7 +3,7 @@ import { calculateCustomTraces } from 'lib/optimization/calculateTraces'
 import { emptyLightCone } from 'lib/optimization/optimizerUtils'
 import { transformComboState } from 'lib/optimization/rotation/comboStateTransform'
 import { StatCalculator } from 'lib/relics/statCalculator'
-import { initializeContextConditionals } from 'lib/simulations/new/workerContextTransform'
+import { initializeContextConditionals } from 'lib/simulations/contextConditionals'
 import DB from 'lib/state/db'
 import { generateConditionalResolverMetadata } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
 import { Form, Teammate } from 'types/form'
