+++ src/lib/scoring/simScoringUtils.ts
@@ -3,8 +3,8 @@ import { OptimizerDisplayData } from 'lib/optimization/bufferPacker'
 import { ComputedStatsArray, ComputedStatsArrayCore, Key, StatToKey } from 'lib/optimization/computedStatsArray'
 import { StatCalculator } from 'lib/relics/statCalculator'
 
-import { SimulationStatUpgrade } from 'lib/simulations/new/scoringUpgrades'
-import { RunStatSimulationsResult, Simulation } from 'lib/simulations/new/statSimulationTypes'
+import { SimulationStatUpgrade } from 'lib/simulations/scoringUpgrades'
+import { RunStatSimulationsResult, Simulation } from 'lib/simulations/statSimulationTypes'
 import { Utils } from 'lib/utils/utils'
 import { Form } from 'types/form'
 import { DBMetadataCharacter, SimulationMetadata } from 'types/metadata'
