+++ src/lib/simulations/optimizerDisplayDataTransform.ts
