+++ src/lib/simulations/orchestrator/benchmarkSimulationOrchestrator.ts
@@ -18,14 +18,14 @@ import {
   SimulationScore,
   spdRollsCap,
 } from 'lib/scoring/simScoringUtils'
-import { generatePartialSimulations } from 'lib/simulations/new/benchmarks/simulateBenchmarkBuild'
-import { generateStatImprovements, SimulationStatUpgrade } from 'lib/simulations/new/scoringUpgrades'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
-import { RunSimulationsParams, RunStatSimulationsResult, Simulation, SimulationRequest, StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
-import { generateFullDefaultForm } from 'lib/simulations/new/utils/benchmarkForm'
-import { applySpeedFlags, calculateTargetSpeedNew } from 'lib/simulations/new/utils/benchmarkSpeedTargets'
-import { runComputeOptimalSimulationWorker } from 'lib/simulations/new/workerPool'
+import { generatePartialSimulations } from 'lib/simulations/benchmarks/simulateBenchmarkBuild'
+import { generateStatImprovements, SimulationStatUpgrade } from 'lib/simulations/scoringUpgrades'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
 import { convertRelicsToSimulation } from 'lib/simulations/statSimulationController'
+import { RunSimulationsParams, RunStatSimulationsResult, Simulation, SimulationRequest, StatSimTypes } from 'lib/simulations/statSimulationTypes'
+import { generateFullDefaultForm } from 'lib/simulations/utils/benchmarkForm'
+import { applySpeedFlags, calculateTargetSpeedNew } from 'lib/simulations/utils/benchmarkSpeedTargets'
+import { runComputeOptimalSimulationWorker } from 'lib/simulations/workerPool'
 import { SimpleCharacter } from 'lib/tabs/tabBenchmarks/UseBenchmarksTabStore'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { ComputeOptimalSimulationWorkerInput, ComputeOptimalSimulationWorkerOutput } from 'lib/worker/computeOptimalSimulationWorkerRunner'
