+++ src/lib/simulations/scoringUpgrades.ts
@@ -1,8 +1,8 @@
 import { MainStatParts, Parts, Stats } from 'lib/constants/constants'
 import { applyScoringFunction, ScoringParams, SimulationFlags } from 'lib/scoring/simScoringUtils'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
-import { RunStatSimulationsResult, Simulation, SimulationRequest } from 'lib/simulations/new/statSimulationTypes'
-import { partsToFilterMapping } from 'lib/simulations/new/utils/benchmarkUtils'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
+import { RunStatSimulationsResult, Simulation, SimulationRequest } from 'lib/simulations/statSimulationTypes'
+import { partsToFilterMapping } from 'lib/simulations/utils/benchmarkUtils'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { Form } from 'types/form'
 import { SimulationMetadata } from 'types/metadata'
