+++ src/lib/simulations/statSimulation.ts
@@ -3,8 +3,8 @@ import { BasicStatsArray, BasicStatsArrayCore } from 'lib/optimization/basicStat
 import { ComputedStatsArray, ComputedStatsArrayCore, Key } from 'lib/optimization/computedStatsArray'
 import { StatCalculator } from 'lib/relics/statCalculator'
 import { cloneSimResult, SimulationFlags } from 'lib/scoring/simScoringUtils'
-import { simulateBuild } from 'lib/simulations/new/simulateBuild'
-import { RunSimulationsParams, RunStatSimulationsResult, Simulation, SimulationRelic, SimulationRelicByPart, StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
+import { simulateBuild } from 'lib/simulations/simulateBuild'
+import { RunSimulationsParams, RunStatSimulationsResult, Simulation, SimulationRelic, SimulationRelicByPart, StatSimTypes } from 'lib/simulations/statSimulationTypes'
 import { precisionRound } from 'lib/utils/mathUtils'
 import { isFlat } from 'lib/utils/statUtils'
 import { Form } from 'types/form'
