+++ src/lib/simulations/statSimulationController.tsx
@@ -9,9 +9,9 @@ import { calculateCurrentlyEquippedRow } from 'lib/optimization/optimizer'
 import { SortOption } from 'lib/optimization/sortOptions'
 import { StatCalculator } from 'lib/relics/statCalculator'
 import { Assets } from 'lib/rendering/assets'
-import { transformOptimizerDisplayData } from 'lib/simulations/new/optimizerDisplayDataTransform'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
-import { Simulation, SimulationRequest, StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
+import { transformOptimizerDisplayData } from 'lib/simulations/optimizerDisplayDataTransform'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
+import { Simulation, SimulationRequest, StatSimTypes } from 'lib/simulations/statSimulationTypes'
 import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
