+++ src/lib/simulations/statSimulationTypes.ts
