+++ src/lib/tabs/tabBenchmarks/BenchmarkResults.tsx
@@ -11,8 +11,8 @@ import { ElementToDamage, SubStats } from 'lib/constants/constants'
 import { toBasicStatsObject } from 'lib/optimization/basicStatsArray'
 import { toComputedStatsObject } from 'lib/optimization/computedStatsArray'
 import { Assets } from 'lib/rendering/assets'
-import { BenchmarkSimulationOrchestrator } from 'lib/simulations/new/orchestrator/benchmarkSimulationOrchestrator'
-import { Simulation } from 'lib/simulations/new/statSimulationTypes'
+import { BenchmarkSimulationOrchestrator } from 'lib/simulations/orchestrator/benchmarkSimulationOrchestrator'
+import { Simulation } from 'lib/simulations/statSimulationTypes'
 import DB from 'lib/state/db'
 import { useBenchmarksTabStore } from 'lib/tabs/tabBenchmarks/UseBenchmarksTabStore'
 import { arrowColor } from 'lib/tabs/tabOptimizer/analysis/StatsDiffCard'
