+++ src/lib/tabs/tabBenchmarks/BenchmarksTab.tsx
@@ -3,7 +3,7 @@ import { Button, Card, Flex, Form as AntDForm, InputNumber, Radio } from 'antd'
 import { OverlayText, showcaseOutline } from 'lib/characterPreview/CharacterPreviewComponents'
 import CharacterModal from 'lib/overlays/modals/CharacterModal'
 import { Assets } from 'lib/rendering/assets'
-import { StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
+import { StatSimTypes } from 'lib/simulations/statSimulationTypes'
 import DB from 'lib/state/db'
 import { BenchmarkResults } from 'lib/tabs/tabBenchmarks/BenchmarkResults'
 import { BenchmarkSetting } from 'lib/tabs/tabBenchmarks/BenchmarkSettings'
