+++ src/lib/tabs/tabOptimizer/analysis/expandedDataPanelController.ts
@@ -7,10 +7,10 @@ import { Buff, ComputedStatsArray, ComputedStatsArrayCore } from 'lib/optimizati
 import { generateContext } from 'lib/optimization/context/calculateContext'
 import { RelicFilters } from 'lib/relics/relicFilters'
 import { aggregateCombatBuffs } from 'lib/simulations/combatBuffsAnalysis'
-import { simulateBuild } from 'lib/simulations/new/simulateBuild'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
-import { Simulation, SimulationRelicByPart, SimulationRequest, StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
+import { simulateBuild } from 'lib/simulations/simulateBuild'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
 import { convertRelicsToSimulation, ornamentSetIndexToName, relicSetIndexToNames } from 'lib/simulations/statSimulationController'
+import { Simulation, SimulationRelicByPart, SimulationRequest, StatSimTypes } from 'lib/simulations/statSimulationTypes'
 import DB from 'lib/state/db'
 import { optimizerFormCache } from 'lib/tabs/tabOptimizer/optimizerForm/OptimizerForm'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
