+++ src/lib/tabs/tabOptimizer/optimizerForm/components/SimulatedBuildsGrid.tsx
@@ -2,8 +2,8 @@ import { CloseOutlined } from '@ant-design/icons'
 import { IRowNode } from 'ag-grid-community'
 import { Empty, Flex, Table, TableColumnsType } from 'antd'
 import { OptimizerDisplayData } from 'lib/optimization/bufferPacker'
-import { Simulation } from 'lib/simulations/new/statSimulationTypes'
 import { deleteStatSimulationBuild, renderDefaultSimulationName } from 'lib/simulations/statSimulationController'
+import { Simulation } from 'lib/simulations/statSimulationTypes'
 import { STAT_SIMULATION_GRID_WIDTH } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
 import { TsUtils } from 'lib/utils/TsUtils'
 import React, { useEffect } from 'react'
