+++ src/lib/worker/computeOptimalSimulationWorker.ts
@@ -1,9 +1,9 @@
 import { Stats, SubStats } from 'lib/constants/constants'
 import { StatToKey } from 'lib/optimization/computedStatsArray'
 import { applyScoringFunction, SimulationResult, substatRollsModifier } from 'lib/scoring/simScoringUtils'
-import { runStatSimulations } from 'lib/simulations/new/statSimulation'
-import { Simulation, StatSimulationTypes } from 'lib/simulations/new/statSimulationTypes'
-import { initializeContextConditionals } from 'lib/simulations/new/workerContextTransform'
+import { initializeContextConditionals } from 'lib/simulations/contextConditionals'
+import { runStatSimulations } from 'lib/simulations/statSimulation'
+import { Simulation, StatSimulationTypes } from 'lib/simulations/statSimulationTypes'
 import { sumArray } from 'lib/utils/mathUtils'
 import { Utils } from 'lib/utils/utils'
 import { ComputeOptimalSimulationWorkerInput, ComputeOptimalSimulationWorkerOutput } from 'lib/worker/computeOptimalSimulationWorkerRunner'
