+++ src/types/store.d.ts
@@ -5,7 +5,7 @@ import { BUFF_TYPE } from 'lib/optimization/buffSource'
 import { Buff } from 'lib/optimization/computedStatsArray'
 import { ColorThemeOverrides } from 'lib/rendering/theme'
 import { ScoringType } from 'lib/scoring/simScoringUtils'
-import { Simulation, StatSimTypes } from 'lib/simulations/new/statSimulationTypes'
+import { Simulation, StatSimTypes } from 'lib/simulations/statSimulationTypes'
 import { ComboState } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
 import { WarpRequest, WarpResult } from 'lib/tabs/tabWarp/warpCalculatorController'
 import { Build, Character, CharacterId } from 'types/character'
